refactor: services and eventqueue
