+++ src/main/scala/ulisse/applications/useCases/StationService.scala
@@ -1,7 +1,6 @@
 package ulisse.applications.useCases
 
-import cats.data.NonEmptyChain
-import ulisse.applications.StationEventQueue
+import ulisse.applications.EventQueues.StationEventQueue
 import ulisse.applications.managers.StationManager
 import ulisse.applications.ports.StationPorts
 import ulisse.entities.Coordinate
@@ -20,7 +19,7 @@ final case class StationService(private val eventQueue: StationEventQueue) exten
     val p = Promise[Either[E, SM]]()
     eventQueue.addCreateStationEvent(stationManager =>
       val updatedMap = stationManager.addStation(station)
-      updateState(p, stationManager, updatedMap)
+      Services.updateManager(p, stationManager, updatedMap, _.stations)
     )
     p.future
 
@@ -28,7 +27,7 @@ final case class StationService(private val eventQueue: StationEventQueue) exten
     val p = Promise[Either[E, SM]]()
     eventQueue.addDeleteStationEvent((stationManager, routeManager, timetableManager) =>
       val updatedMap = stationManager.removeStation(station)
-      (updateState(p, stationManager, updatedMap), routeManager, timetableManager)
+      (Services.updateManager(p, stationManager, updatedMap, _.stations), routeManager, timetableManager)
     )
     p.future
 
@@ -36,7 +35,7 @@ final case class StationService(private val eventQueue: StationEventQueue) exten
     val p = Promise[Either[E, SM]]()
     eventQueue.addUpdateStationEvent((stationManager, routeManager, timetableManager) =>
       val updatedMap = stationManager.removeStation(oldStation).flatMap(_.addStation(newStation))
-      (updateState(p, stationManager, updatedMap), routeManager, timetableManager)
+      (Services.updateManager(p, stationManager, updatedMap, _.stations), routeManager, timetableManager)
     )
     p.future
 
@@ -47,12 +46,3 @@ final case class StationService(private val eventQueue: StationEventQueue) exten
       p.success(station)
     )
     p.future
-
-  private def updateState(
-      p: Promise[Either[E, SM]],
-      stationManager: StationManager,
-      updatedMap: Either[NonEmptyChain[StationManager.Error], StationManager]
-  ) =
-    updatedMap match
-      case Left(value: E) => p.success(Left(value)); stationManager
-      case Right(value)   => p.success(Right(value.stations)); value
